+++ cmd/hidden_lake/applications/filer/_daemon/status_hlf.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-watch -c SYSTEMD_COLORS=1 systemctl status -o cat hidden_lake_filer.service