+++ cmd/hidden_lake/applications/filesharer/_daemon/journal_hlf.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+journalctl -eu hidden_lake_filesharer.service